fix: Round Value of BigDecimal

+++ app/src/main/java/com/arduia/expense/data/backup/ExpenseBackupSheet.kt
@@ -4,6 +4,7 @@ import com.arduia.backup.BackupSheet
 import com.arduia.backup.BackupSource
 import com.arduia.expense.data.local.ExpenseEnt
 import com.arduia.expense.domain.Amount
+import java.math.BigDecimal
 import java.util.*
 
 class ExpenseBackupSheet( source: BackupSource<ExpenseEnt>): BackupSheet<ExpenseEnt>(source){
@@ -20,7 +21,7 @@ class ExpenseBackupSheet( source: BackupSource<ExpenseEnt>): BackupSheet<Expense
         val note = row[FIELD_NOTE]
         val date = row[FIELD_DATE]?.toLong() ?: 0
 
-            return ExpenseEnt(0, name, Amount.createFromActual(amount), category, note, date, Date().time)
+            return ExpenseEnt(0, name, Amount.createFromActual(BigDecimal(amount.toDouble())), category, note, date, Date().time)
     }
 
     override fun mapToSheetDataFromEntity(item: ExpenseEnt): Map<String, String> {
