+++ app/src/main/java/com/arduia/expense/domain/DataStoreExchangeRate.kt
@@ -2,6 +2,6 @@ package com.arduia.expense.domain
 
 import com.arduia.currencystore.Rate
 
-object DataStoreExchangeRate: Rate<Int> {
-    override fun getRate(): Int = 100
+object DataStoreExchangeRate: Rate<Long> {
+    override fun getRate(): Long = 100L
 }
