+++ app/src/main/java/com/arduia/expense/ui/entry/ExpenseEntryViewModel.kt
@@ -16,6 +16,7 @@ import kotlinx.coroutines.flow.single
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import java.lang.Exception
+import java.math.BigDecimal
 import java.util.*
 
 class ExpenseEntryViewModel @ViewModelInject constructor(
@@ -137,7 +138,7 @@ class ExpenseEntryViewModel @ViewModelInject constructor(
     ) = ExpenseEnt(
         expenseId = vto.id,
         name = vto.name,
-        amount = Amount.createFromActual(vto.amount.toFloat()),
+        amount = Amount.createFromActual(BigDecimal(vto.amount)),
         note = vto.note,
         category = vto.category,
         createdDate = createdDate ?: Date().time,
